Enable READONLY file server.

+++ docker-entrypoint.sh
@@ -4,8 +4,10 @@ set -e
 if [ "$1" = 'nginx' ]; then
   python /build-conf-files.py
   service supervisor start
-  rm -rf /database_attachments/_tmp
-  mkdir /database_attachments/_tmp
+  if [ x"$READ_ONLY_FILESERVER" = 'x' ]; then
+    rm -rf /database_attachments/_tmp
+    mkdir /database_attachments/_tmp
+  fi
 fi
 
 exec "$@"
